+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and adheres to [Semantic Versioning](https://semver.org/).
 
 - fix optimized N+1 problem in API test -
   GetsQuarterHourlyAggregatesByLocationLast
+- added dapper queries for getting chunk interval information (WIP)
+- fix MeasurementDeletionJobReactorTest now uses chunk interval information to
+  determine which measurement chunks will be removed from db
 - updated HtmlSanitizer to version 9.0.892
 - fix ReadLastByMeasurementLocationIds now returns latest measurement info for
   all meters that have measurements